+++ src/testing/index.ts
@@ -1,7 +1,10 @@
-export { createJestPuppeteerEnvironment } from './jest/jest-27-and-under/jest-environment';
-export { jestPreprocessor } from './jest/jest-27-and-under/jest-preprocessor';
-export { createTestRunner } from './jest/jest-27-and-under/jest-runner';
-export { jestSetupTestFramework } from './jest/jest-27-and-under/jest-setup-test-framework';
+export {
+  getCreateJestPuppeteerEnvironment,
+  getCreateJestTestRunner,
+  getJestPreprocessor,
+  getJestPreset,
+  getJestSetupTestFramework,
+} from './jest/jest-stencil-connector';
 export {
   mockFetch,
   MockHeaders,